pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secbot"
version = "1.10.1b2"
description = "Hackbot: AI-powered automated penetration testing robot"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "赵明俊", email = "wisewater5419@gmail.com"}
]
keywords = ["ai", "agent", "security", "automation", "llm", "langchain"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Security",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-core>=0.1.23",
    "langchain-experimental>=0.0.50",
    "langchain-ollama>=0.1.0",
    "langchain-openai>=0.2.0",
    "aiohttp>=3.9.1",
    "httpx>=0.26.0",
    "typer>=0.9.0",
    "rich>=13.7.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",
    "selenium>=4.17.0",
    "playwright>=1.41.0",
    "sqlalchemy>=2.0.25",
    "sqlite-vec>=0.1.0; sys_platform != 'win32'",
    "sqlite-vss>=0.1.0; sys_platform != 'win32'",
    "psutil>=5.9.0",
    "distro>=1.8.0",
    "loguru>=0.7.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "python-multipart>=0.0.6",
    "websockets>=12.0",
    "tiktoken>=0.5.2",
    "keyring>=25.0.0",
    "prompt_toolkit>=3.0.0",
    "paramiko>=3.0.0",
    "setuptools>=68.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sse-starlette>=1.8.0",
    "langgraph>=0.2.0",
    "ddgs>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.3",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
cli = ["keyring>=25.0.0"]
anthropic = ["langchain-anthropic>=0.3.0"]
google = ["langchain-google-genai>=2.0.0"]
all-providers = [
    "langchain-anthropic>=0.3.0",
    "langchain-google-genai>=2.0.0",
]
exploit-tools = [
    "pymetasploit3>=1.0.0",
]

[project.scripts]
hackbot = "secbot_cli.cli:app"
secbot = "secbot_cli.cli:app"
secbot-cli = "secbot_cli.cli:app"

# 工具扩展 entry point：第三方包在其 pyproject.toml 中声明即可注册
# 文档见 tools/registry.py 或 docs/TOOL_EXTENSION.md

[project.urls]
Homepage = "https://github.com/iammm0/hackbot"
Documentation = "https://github.com/iammm0/hackbot/docs"
Repository = "https://github.com/iammm0/hackbot"
Issues = "https://github.com/iammm0/hackbot/issues"

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
tag_format = "v{version}"
changelog_file = "CHANGELOG.md"
commit_parser = "angular"
build_command = ""
commit_message = "chore(release): {version} [skip ci]"
upload_to_pypi = false
upload_to_release = true

[tool.semantic_release.branches.main]
match = "^(main)$"
prerelease = false

[tool.semantic_release.branches.beta]
match = "^(beta)$"
prerelease = true
prerelease_token = "beta"

[tool.setuptools.packages.find]
where = ["."]
include = [
    "hackbot_config*",
    "router*",
    "secbot_agent*",
    "secbot_cli*",
    "system*",
    "tools*",
    "utils*",
]
exclude = ["tests*"]

[tool.setuptools.package-data]
"secbot_agent.prompts" = ["templates/*.yaml", "templates/*.yml"]
"secbot_agent.skills" = ["base/*/*.md"]
"*" = ["*.yaml", "*.yml", "*.txt", "*.md", "env.example"]

[tool.setuptools.exclude-package-data]
"*" = ["*.env", "*.env.*", "data/*", "logs/*"]

[tool.black]
line-length = 120
target-version = ['py310', 'py311', 'py312']

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "sys_platform == 'win32' and platform_machine == 'AMD64'",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
